#SPD-1628: adding changes for refactoring
